# app/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: RDF Graphs Sharing/Exhange Server API Specifications
  description: "This is a RDF Graph Exhange Server REST API specifications to provide\
    \ users to access (CRUD - create, retrieve, update, and delete) RDF Graph-compliant\
    \ RDF resources.  - Each user's RDF graph (RDF Graph-compliant) can be shared\
    \ with other users to enable Federated SPARQL query if the aggregated user roles\
    \ and permissions over the RDF graphs passing the Access Control validation. Otherwise,\
    \ the access will be denied. - The create operation can accept data source formats\
    \ JSON-LD, RDF/XML, N3, or Turtle format for upload/create. - It is up to the\
    \ implementation to support multiple data source formats. - The RDF Graph Exchange\
    \ reference implementation will only support one of JSON-LD, N3, or Turtle formats."
  version: 0.1.3
servers:
- url: https://virtserver.swaggerhub.com/RS-API/RDF-Graph-Exchange/0.1.3
  description: SwaggerHub API Auto Mocking
paths:
...
  /api/access-control/create:
    post:
      tags:
      - access-control
      summary: upload and create a new RDF graph upload RDF graph data source
      operationId: access_create
      parameters:
      - name: access_token
        in: query
        description: Authorization access token string
        required: true
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/AccessToken'
      - name: userID
        in: query
        description: the ID of the organization of the client application
        required: true
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: UCO-CASE-EU-CC-1234567890
          default: userUUID-unknown
      - name: graphURI
        in: query
        required: true
        style: form
        explode: true
        schema:
          maxLength: 128
          minLength: 6
          type: string
          example: <http://uco.example.org/bookStore>
          default: <http://uco.example.org/bookstore>
      - name: organizationID
        in: query
        description: the organization of the client application
        required: false
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: uco-eu-cyber-123
          default: organizationID-unknown
      - name: accessControl
        in: query
        description: the access control record (object)
        required: true
        style: form
        explode: true
        schema:
          type: string
          nullable: false
          example: JSON-LD
          default: JSON-LD
          enum:
          - JSON-LD
          - TRIG
          - RDF/XML
          - N3
          - TURTLE
      - name: reference
        in: query
        description: the client applications' reference to this batch of RDF data
          to be uploaded and it is not used by the Server.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: cs0123-b45678-2021-11-25-001
          default: n/a
      requestBody:
        description: upload RDF data source
        content:
          application/text:
            schema:
              $ref: '#/components/schemas/SourceDataType'
      responses:
        "200":
          description: successful operation
          content:
            application/text:
              schema:
                type: string
                description: Results in text format
                default: "{}"
                x-content-type: application/text
        "400":
          description: "Bad request, exact description of error and any successful\
            \ results to be in content"
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "403":
          description: No permission to access object(s)
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "404":
          description: Specified object(s) not found
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "500":
          description: Server was not able to process request
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
      security: []
      x-openapi-router-controller: swagger_server.controllers.access_control_controller
  /api/access-control/query:
    post:
      tags:
      - access-control
      summary: Either SPARQL v1.1. RDF (Federated) Query or GraphQL (Federated) Query
        from multiple data sources
      operationId: access_query
      parameters:
      - name: access_token
        in: query
        description: Authorization access token string
        required: true
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/AccessToken'
      - name: userID
        in: query
        description: the ID of the organization of the client application
        required: true
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: UCO-CASE-EU-CC-1234567890
          default: userUUID-unknown
      - name: graphURIList
        in: query
        required: true
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/GraphListType'
      - name: organizationID
        in: query
        description: the organization of the client application
        required: false
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: uco-eu-cyber-123
          default: organizationID-unknown
      - name: reference
        in: query
        description: the client applications' reference to this batch of RDF data
          to be uploaded and it is not used by the Server.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: cs0123-b45678-2021-11-25-001
          default: n/a
      requestBody:
        description: More specifics about what RDF data to delete using SPARQL 'DELETE
          DATA' statement. Note the Client can only delete the original Graph data
          sources being uploaded by the same User ID and Graph name. These graphs
          uploaded by others will not be deleted by the request as the protection
          by the ownership of those other graphs.
        content:
          application/text:
            schema:
              $ref: '#/components/schemas/SourceDataType'
      responses:
        "200":
          description: successful operation
          content:
            application/text:
              schema:
                type: string
                description: Results in text format
                default: "{}"
                x-content-type: application/text
        "400":
          description: "Bad request, exact description of error and any successful\
            \ results to be in content"
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "403":
          description: No permission to access object(s)
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "404":
          description: Specified object(s) not found
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "500":
          description: Server was not able to process request
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
      security: []
      x-openapi-router-controller: swagger_server.controllers.access_control_controller
  /api/access-control/update:
    post:
      tags:
      - access-control
      summary: update RDF graph data source
      operationId: access_update
      parameters:
      - name: access_token
        in: query
        description: Authorization access token string
        required: true
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/AccessToken'
      - name: userID
        in: query
        description: the ID of the organization of the client application
        required: true
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: UCO-CASE-EU-CC-1234567890
          default: userUUID-unknown
      - name: graphURI
        in: query
        required: true
        style: form
        explode: true
        schema:
          maxLength: 128
          minLength: 6
          type: string
          example: <http://uco.example.org/bookStore>
          default: <http://uco.example.org/bookstore>
      - name: organizationID
        in: query
        description: the organization of the client application
        required: false
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: uco-eu-cyber-123
          default: organizationID-unknown
      - name: reference
        in: query
        description: the client applications' reference to this batch of RDF data
          to be uploaded and it is not used by the Server.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: cs0123-b45678-2021-11-25-001
          default: n/a
      requestBody:
        description: SPARQL v1.1 (see W3C SPARQL v1.1. Update Specification - https://www.w3.org/TR/sparql11-update/
          ) as the specification syntax of the payload SPARQL Update statements.
        content:
          application/text:
            schema:
              $ref: '#/components/schemas/SourceDataType'
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Results'
        "400":
          description: "Bad request, exact description of error and any successful\
            \ results to be in content"
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "403":
          description: No permission to access object(s)
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "404":
          description: Specified object(s) not found
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "500":
          description: Server was not able to process request
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
      security: []
      x-openapi-router-controller: swagger_server.controllers.access_control_controller
  /api/access-control/delete:
    post:
      tags:
      - access-control
      summary: Delete RDF graph data source
      operationId: access_delete
      parameters:
      - name: access_token
        in: query
        description: Authorization access token string
        required: true
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/AccessToken'
      - name: userID
        in: query
        description: the ID of the organization of the client application
        required: true
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: UCO-CASE-EU-CC-1234567890
          default: userUUID-unknown
      - name: graphURI
        in: query
        required: true
        style: form
        explode: true
        schema:
          maxLength: 128
          minLength: 6
          type: string
          example: <http://uco.example.org/bookStore>
          default: <http://uco.example.org/bookstore>
      - name: organizationID
        in: query
        description: the organization of the client application
        required: false
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: uco-eu-cyber-123
          default: organizationID-unknown
      - name: reference
        in: query
        description: the client applications' reference to this batch of RDF data
          to be uploaded and it is not used by the Server.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: cs0123-b45678-2021-11-25-001
          default: n/a
      requestBody:
        description: More specifics about what RDF data to delete using SPARQL 'DELETE
          DATA' statement. Note the Client can only delete the original Graph data
          sources being uploaded by the same User ID and Graph name. These graphs
          uploaded by others will not be deleted by the request as the protection
          by the ownership of those other graphs.
        content:
          application/text:
            schema:
              $ref: '#/components/schemas/SourceDataType'
      responses:
        "200":
          description: successful operation
          content:
            application/text:
              schema:
                type: string
                description: Results in text format
                default: "{}"
                x-content-type: application/text
        "400":
          description: "Bad request, exact description of error and any successful\
            \ results to be in content"
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "403":
          description: No permission to access object(s)
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "404":
          description: Specified object(s) not found
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "500":
          description: Server was not able to process request
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
      security: []
      x-openapi-router-controller: swagger_server.controllers.access_control_controller
...
  /graphql:
    post:
      tags:
      - graphql
      summary: GraphQL REST API
      operationId: graphql
      parameters:
      - name: access_token
        in: query
        description: Authorization access token string
        required: true
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/AccessToken'
      - name: userID
        in: query
        description: the ID of the organization of the client application
        required: true
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: UCO-CASE-EU-CC-1234567890
          default: userUUID-unknown
      - name: graphList
        in: query
        description: Either SPARQL or GraphQL query is supported
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/GraphListType'
      - name: organizationID
        in: query
        description: the organization of the client application
        required: false
        style: form
        explode: true
        schema:
          maxLength: 64
          minLength: 6
          type: string
          example: uco-eu-cyber-123
          default: organizationID-unknown
      - name: reference
        in: query
        description: the client applications' reference to this batch of RDF data
          to be uploaded and it is not used by the Server.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: cs0123-b45678-2021-11-25-001
          default: n/a
      requestBody:
        description: GraphQL query.
        content:
          application/text:
            schema:
              description: "GraphQL (Federated) Query Statement as the body content.\
                \ If the body content is empty, then the entire RDF Graph (parameter\
                \ graphURI) will be used to query against the specific RDF graph in\
                \ the Graph Exchange Server. Only the users with authorized access\
                \ API token can the operation."
      responses:
        "200":
          description: successful operation
          content:
            application/text:
              schema:
                type: string
                description: Results in text format
                default: "{}"
                x-content-type: application/text
        "400":
          description: "Bad request, exact description of error and any successful\
            \ results to be in content"
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "403":
          description: No permission to access object(s)
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "404":
          description: Specified object(s) not found
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
        "500":
          description: Server was not able to process request
          content:
            application/text:
              schema:
                $ref: '#/components/schemas/Errors'
      security: []
      x-openapi-router-controller: swagger_server.controllers.graphql_controller
...
components:
  schemas:
    AccessToken:
      type: string
      description: "Authorization Access Token in OAuth 2.0 and there are a few different\
        \ format. This schema assumes the existence of Authorization Access Server\
        \ such as OAuth 2.0 Servers besides the Server so that the Clients / Applications\
        \ / Users can request Access Token (short-lived or long-lived) to access this\
        \ the API services. And, the API Interface here will not prescribe the formats.\
        \ One of the common formats for Access Token is Json Web Token (JWT). A JWT\
        \ is represented as a sequence of base64url encoded values that are separated\
        \ by period characters with three parts as <Header>.<Payload>.<Signature>\
        \ - see the example value. And, the Interface will not prescribe the specific\
        \ format here."
      example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vdHJ1c3R5YXBwLmNvbS8iLCJleHAiOjEzMDA4MTkzODAsInN1YiI6InVzZXJzLzg5ODM0NjIiLCJzY29wZSI6InNlbGYgYXBpL2J1eSJ9.43DXvhrwMGeLLlP4P4izjgsBB2yrpo82oiUPhADakLs
    AccessPermissionType:
      type: string
      description: "the choice of the RDF Graph access permission: READ-ONLY, WRITE,\
        \ and READ-WRITE. The WRITE permission will allow a user or group to 'Add,\
        \ Update, Delete' RDF resources in the RDF Graph."
      nullable: false
      example: READ-ONLY
      default: READ-ONLY
      enum:
      - READ-ONLY
      - WRITE
      - READ-WRITE
...
    Errors:
      required:
      - code
      - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    UserType:
      required:
      - userID
      properties:
        userUUID:
          maxLength: 64
          type: string
          description: the UUID (v4.0) of an organization to be generated by the RDF
            Graph Exchange Server.
          nullable: true
          example: 6ff13385-070b-46d0-9180-c91b93d0a81b
          default: abcdabcd-0000-1234-1234-abcdefabcdef
        userID:
          type: string
          description: a non-UUID (short) organization UID and it could be the same
            as user's email.
          nullable: false
          example: uco-user-0123456
          default: uco-user-0123456
        userFirstName:
          type: string
          nullable: true
          example: Jane
        userLastName:
          type: string
          nullable: true
          example: Smith
        email:
          type: string
          nullable: true
          example: user12345678@uco.acme.org
        affiliatedOrganization:
          $ref: '#/components/schemas/OrganizationType'
        dateStart:
          type: string
          description: starting date-time the user allowed to access the RDF Graph
            Exchange Server
          format: date-time
        dateEnd:
          type: string
          description: expiration date-time the user allowed to access the RDF Graph
            Exchange Server
          format: date-time
          example: 2022-11-18T19:23:38.04Z
        graphAccessControlList:
          $ref: '#/components/schemas/GraphAccessControlListType'
      description: User information
...
    OrganizationType:
      required:
      - organizationID
      properties:
        organizationUUID:
          maxLength: 64
          type: string
          description: the UUID (v4.0) of an organization to be generated by the RDF
            Graph Exchange Server.
          nullable: true
          example: 6ff13385-070b-46d0-9180-c91b93d0a81b
          default: abcdabcd-1234-1234-1234-abcdefabcdef
        organizationID:
          maxLength: 64
          minLength: 6
          type: string
          description: a non-UUID ()short) organization UID
          nullable: false
          example: US-CYBER-ORG-01234
          default: organizationID-unknown
        organizationName:
          maxLength: 64
          minLength: 6
          type: string
          nullable: true
          example: ACME Ontology Inc.
          default: organizationName-unknown
        organizationWeb:
          maxLength: 128
          type: string
          nullable: true
          example: https://github.com/ucoProject/UCO
          default: organizationWeb-unknown
      description: user's Organization information
...
    GroupType:
      required:
      - groupID
      properties:
        groupID:
          maxLength: 64
          minLength: 6
          type: string
          nullable: false
          example: uco-eu-cyber-123
          default: groupID-unknown
        groupName:
          maxLength: 64
          type: string
          example: EU Cyber Ontology Group
          default: groupName-unknown
        remark:
          maxLength: 64
          type: string
          nullable: true
          example: The EU Cyber Ontology Group - studying the adoption of RDF Graph
            in cyber domain.
      description: user's group / role ID and name
...
    GraphType:
      required:
      - graphURI
      - owner
      properties:
        organizationUUID:
          maxLength: 64
          type: string
          description: the UUID (v4.0) of an organization to be generated by the RDF
            Graph Exchange Server.
          nullable: true
          example: 6ff13385-070b-46d0-9180-c91b93d0a81b
          default: abcdabcd-1234-1234-1234-abcdefabcdef
        graphURI:
          maxLength: 128
          minLength: 6
          type: string
          description: the URI of the RDF Graph data sources
          example: <http://uco.example.org/bookStore>
          default: <http://uco.example.org/bookstore>
        sharable:
          type: boolean
          description: whether to allow the RDF graph to be shared other than the
            OWNER (the user who uploaded/created the RDF Graph)
          default: false
        owner:
          $ref: '#/components/schemas/UserType'
        organization:
          $ref: '#/components/schemas/OrganizationType'
        dateCreated:
          type: string
          description: date the RDF graph being uploaded/created by the user. Note
            this will auto-filled by the RDF Graph Exchange Server if not provided
          format: date-time
          example: 2021-11-14T19:23:38.04Z
        dateExpired:
          type: string
          description: "date the RDF graph will be expired and removed automatically\
            \ or not available for sharing. But, the owner can still access with RUD\
            \ (Query, Update, Delete) operations. Note this will auto-default with,\
            \ e.g., 90 days to expire, if not provided by the user."
          format: date-time
          example: 2022-11-14T19:23:38.04Z
      description: the RDF Graph that the user uploaded/created
    GraphListType:
      type: array
      description: Graph list
      example:
      - <http://uco.example.org/bookStore>
      - <http://uco.example.org/student>
      items:
        $ref: '#/components/schemas/GraphType'
    GraphAccessGroupType:
      required:
      - accessGroup
      - accessPermission
      - graphURI
      properties:
        graphURI:
          maxLength: 128
          minLength: 6
          type: string
          description: the URI of the RDF Graph data sources
          example: <http://uco.example.org/bookStore>
          default: <http://uco.example.org/bookstore>
        accessGroup:
          $ref: '#/components/schemas/GroupType'
        accessPermission:
          $ref: '#/components/schemas/AccessPermissionType'
      description: "the access control object for a RDF Graph resource as the ternery\
        \ relationship of the access control permission, the RDF Graph (URI), and\
        \ the specific Group."
    GraphAccessControlListType:
      type: array
      description: the list of GraphAccessControl objects
      items:
        $ref: '#/components/schemas/GraphAccessGroupType'
...
    Results:
      type: string
      description: JSON as Query results
      example: "{}"
...
    SourceDataType:
      type: string
      example: " @prefix dc: <http://purl.org/dc/elements/1.1/> . @prefix ns: <http://example.org/ns#>\
        \ . <http://example/book2> ns:price 42 . <http://example/book2> dc:title \"\
        David Copperfield\" . <http://example/book2> dc:creator \"Edmund Wells\" ."
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      x-basicInfoFunc: swagger_server.controllers.authorization_controller.check_BasicAuth
    BearerAuth:
      type: http
      scheme: bearer
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_BearerAuth
    ApiKeyAuth:
      type: apiKey
      name: X-API-Key
      in: header
      x-apikeyInfoFunc: swagger_server.controllers.authorization_controller.check_ApiKeyAuth
    OpenID:
      type: openIdConnect
      openIdConnectUrl: https://example.com/.well-known/openid-configuration
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://example.com/oauth/authorize
          tokenUrl: https://example.com/oauth/token
          scopes:
            read: Grants read access
            write: Grants write access
            admin: Grants access to admin operations
      x-tokenInfoFunc: swagger_server.controllers.authorization_controller.check_OAuth2
      x-scopeValidateFunc: swagger_server.controllers.authorization_controller.validate_scope_OAuth2